Separate Map component of mapresults into it's own component
