+++ components/containers/MapLocation.js
@@ -1,9 +1,11 @@
 import React, {Component} from 'react';
 import { withScriptjs, withGoogleMap, GoogleMap, Marker, InfoWindow } from 'react-google-maps';
 
-export default class MapLocations extends Component {
+export default class MapLocation extends Component {
 	constructor( props ) {
-		super( props)
+		super( props )
+		//Checks if the isSelected prop is true or false. Set to false by default. So if the name is clicked, 
+		//the marker should appear in map results.
 
 		this.state = ( {
 			isOpen: props.isSelected || false,
@@ -19,8 +21,7 @@ export default class MapLocations extends Component {
 	// }
  
 	_handleOnClick = () => {
-	this.setState({ isOpen: true });
-		 
+		this.setState({ isOpen: true }); 
 	}
 
 	_handleOnCloseClick = () => {
@@ -43,12 +44,13 @@ export default class MapLocations extends Component {
 
 					{this.state.isOpen && 
 						<InfoWindow onCloseClick={this._handleOnCloseClick}>
-						<div className="marker_label_text">
-						 <span className="name_marker">{ this.props.name }</span>
-						 <img className="marker_img" src={`${ this.props.imageSrc }`} />
-						 <span className="city_state_marker">{ this.props.city}, {this.props.state}, {this.props.zipcode} </span>
-						</div>
-					</InfoWindow> }
+							<div className="marker_label_text">
+							 <span className="name_marker">{ this.props.name }</span>
+							 <img className="marker_img" src={`${ this.props.imageSrc }`} />
+							 <span className="city_state_marker">{ this.props.city}, {this.props.state}, {this.props.zipcode} </span>
+							</div>
+						</InfoWindow> 
+					}
 
 				</Marker>
 			</div>
